reuse VectorDatasetInfo to download vectors for VectorDataset
